🚚 Rename: public 위치 변경에 따른 svg import 경로 변경
